Refactor files from Movies Detail
